Modified Travis, clean_install and documentation scripts:
  - Documentation is created only in master branch.
  - Suppressed build and install messages.
  - Increased custom message verbosity.
  - Fixed documentation issues. (install required packages and remove
    obsolete folders)

(closes #164) [ci skip]

+++ rapp_scripts/documentation/update_rapp-project.github.io.sh
@@ -23,38 +23,50 @@ TEMP_PLATFORM_SCRIPTS="${TEMP_PLATFORM}/rapp_scripts/documentation"
 TEMP_PAGES="${HOME}/rapp_temp/rapp-platform-pages"
 DOCS="${HOME}/rapp_platform_files/documentation"
 
-if [ -d ${TEMP_PATH} ]; then
-  rm -rf ${TEMP_PATH}
+if [ "$TRAVIS_BRACH" == "master" ]; then
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Updating Github Pages\e[0m"
+
+  if [ -d ${TEMP_PATH} ]; then
+    rm -rf ${TEMP_PATH}
+  fi
+  mkdir -p ${TEMP_PATH}
+
+  cd ${TEMP_PATH}
+
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Cloning Repository\e[0m"
+  git clone https://github.com/rapp-project/rapp-platform.git &> /dev/null
+  cp -r rapp-platform rapp-platform-pages
+
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Creating Documentation\e[0m"
+  cd ${TEMP_PLATFORM_SCRIPTS}
+  ./create_documentation.sh &> /dev/null
+
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Creating Test Documentation HTML hyperlinks\e[0m"
+
+  cd ${TEMP_PAGES}
+  git checkout gh-pages &> /dev/null
+
+  for file in ${DOCS}/*; do
+    if [ -d $(basename $file) ]; then
+      rm -rf file
+    fi
+  done
+  mv ${DOCS}/* .
+
+  ./create_test_hyperlinks.py
+
+  cat documentation1.html > documentation.html
+  cat documentation2.html >> documentation.html
+  cat documentation3.html >> documentation.html
+
+  git config user.name "Travis CI"
+  git config user.email "etsardou@iti.gr"
+  git add .
+  git commit -m "Deploy to GitHub Pages"
+  git push --force --quiet https://${GH_TOKEN}@github.com/rapp-project/rapp-platform gh-pages &> /dev/null
+
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Finished Cleanly.\e[0m"
+
+else
+  echo -e "\e[1m\e[103m\e[31m[RAPP] Skipping Github Pages Update, branch: ${TRAVIS_BRANCH}\e[0m"
 fi
-mkdir -p ${TEMP_PATH}
-
-cd ${TEMP_PATH}
-
-echo -e "\e[1m\e[103m\e[31m[RAPP] Cloning Repository\e[0m"
-git clone https://github.com/rapp-project/rapp-platform.git &> /dev/null
-cp -r rapp-platform rapp-platform-pages
-
-echo -e "\e[1m\e[103m\e[31m[RAPP] Creating Documentation\e[0m"
-cd ${TEMP_PLATFORM_SCRIPTS}
-./create_documentation.sh &> /dev/null
-
-echo -e "\e[1m\e[103m\e[31m[RAPP] Creating Test Documentation HTML hyperlinks\e[0m"
-
-cd ${TEMP_PAGES}
-git checkout gh-pages &> /dev/null
-
-mv ${DOCS}/* .
-
-./create_test_hyperlinks.py
-
-cat documentation1.html > documentation.html
-cat documentation2.html >> documentation.html
-cat documentation3.html >> documentation.html
-
-git config user.name "Travis CI"
-git config user.email "etsardou@iti.gr"
-git add .
-git commit -m "Deploy to GitHub Pages"
-git push --force --quiet https://${GH_TOKEN}@github.com/rapp-project/rapp-platform gh-pages &> /dev/null
-
-echo -e "\e[1m\e[103m\e[31m[RAPP] Finished Cleanly.\e[0m"
